03-bridge/requirements.txt
influxdb==5.2.1
paho-mqtt==1.4.0
